# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CGU Dcard Analysis API",
        "description": "CGU Dcard Analysis API",
        "contact": {
            "email": "yuyuchia7423@gmail.com"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Localhost"
        },
        {
            "url": "https://dcard-analysis-laravel-develop-fdqsyjapma-de.a.run.app/",
            "description": "測試區主機"
        },
        {
            "url": "https://dcard-analysis-laravel-fdqsyjapma-de.a.run.app/",
            "description": "正式區主機"
        }
    ],
    "paths": {
...
        "/api/v2/dcard": {
            "get": {
                "tags": [
                    "Dcard"
                ],
                "summary": "Get All Dcards",
                "description": "Get All Dcards",
                "operationId": "dcards",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/v2/searchDcard": {
            "get": {
                "tags": [
                    "Dcard"
                ],
                "summary": "Search Dcards",
                "description": "Search Dcards",
                "operationId": "searchDcards",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/v2/dcard/{id}": {
            "get": {
                "tags": [
                    "Dcard"
                ],
                "summary": "Get Dcard",
                "description": "Get Dcard",
                "operationId": "dcard",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/v2/date/{date1}/{date2}": {
            "get": {
                "tags": [
                    "Dcard"
                ],
                "summary": "Get Date Between Dcards",
                "description": "Get Date Between Dcards",
                "operationId": "dateBetween",
                "parameters": [
                    {
                        "name": "date1",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date2",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/v2/date/{type}": {
            "get": {
                "tags": [
                    "Dcard"
                ],
                "summary": "Get Date Dcards",
                "description": "Get Date Dcards",
                "operationId": "date",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/v2/maxScore/{date1}/{date2}": {
            "get": {
                "tags": [
                    "Dcard"
                ],
                "summary": "Get Max SA_Score",
                "description": "Get Max SA_Score",
                "operationId": "maxScore",
                "parameters": [
                    {
                        "name": "date1",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date2",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/v2/minScore/{date1}/{date2}": {
            "get": {
                "tags": [
                    "Dcard"
                ],
                "summary": "Get Min SA_Score",
                "description": "Get Min SA_Score",
                "operationId": "minScore",
                "parameters": [
                    {
                        "name": "date1",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date2",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    ]
}